Fix #4 Change createWorkspaceEdit to computeCommandEdits
When the service returns a WorkspaceEdit, it is indirectly enforcing
the use of regular changes instead of versioned changes. If a
language server is okay with having versioned changes returned back
to the client, it should be able to create such changes. We should
just return a TextEdit[] back to the client instead of a
WorkspaceEdit.

Signed-off-by: Remy Suen <remy.suen@gmail.com>