Fix #27 Change how the example is built with npm 5
As npm 5 uses symbolic links for local installs, the output of the
example will not actually work because modules cannot be resolved at
runtime.

To fix this, we run `npm pack` so that the language service will be
bundled up. Then we have the example install that tarball instead of
using a local install. This ensures that a physical directory will
exist in example/node_modules/dockerfile-language-service instead of
a symbolic link. The modules will then be resolvable at runtime.

Signed-off-by: Remy Suen <remy.suen@gmail.com>