Fix #32 Warn if JSON written with single quotes
JSON requires strings to be enclosed in double quotes. If an
instruction's argments look like JSON but is written with single
quotes, create a warning to indicate that they should be written with
double quotes instead. It is possible to change this rule to flag the
issue as an error or to have it ignored.

Signed-off-by: Remy Suen <remy.suen@gmail.com>