Improve CHANGELOG's readability
The mention of the Capabilities interface for release 0.0.3 should
only mention the updated properties. As it is in the changelog's
section about additions, there should not be any ambiguity over the
fact that the snippetSupport property is not listed in the interface
definition. The entry has also been updated with a link to the
originating issue as an aid.

Signed-off-by: Remy Suen <remy.suen@gmail.com>