Upgrade Mocha dependency to 5.x
Mocha 3.x dependended on Growl 1.9.2 which had a security
vulnerability. Hence, we are upgrading our dependency to Mocha to
ensure that this security vulnerability is no longer exposed.

Note however that as Mocha is purely used for the testing of
dockerfile-language-service. We have no reason to believe that
dockerfile-language-service itself was affected by this security
vulnerability in Growl.

Signed-off-by: Remy Suen <remy.suen@gmail.com>