Upgrade example's depency of lodash from 4.17.5 to 4.17.11
Security vulnerability CVE-2018-16487 was found in versions of lodash
before 4.17.11. As included example used this package, it was
necessary to upgrade the dependency to prevent others using the
example as a reference from being affected.

All users that used the example as a reference are encouraged to check
their dependencies and update lodash if necessary.

Signed-off-by: Remy Suen <remy.suen@gmail.com>