Fix #120 Correct link range if image has a tag and digest
This was fixed in 0.11.0 but no regression test was added so one will be
added now.

CHANGELOG.md is also retroactively being corrected to include this as
one of the fixes included in 0.11.0. Note that the Git tag itself of
course will not change and will continue to point at
a7e1c23255004f7273062390d3427f76979efe9d which was what was used to
release 0.11.0 to npm.

Signed-off-by: Remy Suen <remy.suen@gmail.com>